Fixes for building under MSVC
- printf.c: change for MSDOS needs to exclude WIN32
- stack.c: explicit cast double to int
- byte_tester/main.c: explicit cast ptr_diff_t to int

+++ stack.c
@@ -1002,7 +1002,7 @@ int stack_calculate(char *in, int *value) {
           }
         }
         else {
-          int priority = _get_op_priority(si[k].value);
+          int priority = _get_op_priority((int)(si[k].value));
 
           b--;
           while (b != -1 && op[b] != SI_OP_LEFT && _get_op_priority(op[b]) >= priority) {
